Refactor mutable variant creation into `AbstractMutableModuleComponentResolveMetadata`
This commit avoids duplication of code by factorizing mutable variant code into the common
superclass for both Ivy and Maven mutable metadata.
