+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/AbstractMutableModuleComponentResolveMetadata.java
@@ -22,13 +22,20 @@ import com.google.common.collect.Maps;
 import org.gradle.api.Action;
 import org.gradle.api.artifacts.DependenciesMetadata;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
+import org.gradle.api.artifacts.VersionConstraint;
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
 import org.gradle.api.internal.artifacts.DefaultModuleVersionIdentifier;
+import org.gradle.api.internal.attributes.ImmutableAttributes;
+import org.gradle.internal.Describables;
+import org.gradle.internal.DisplayName;
 import org.gradle.internal.component.external.descriptor.Configuration;
+import org.gradle.internal.component.model.ComponentArtifactMetadata;
+import org.gradle.internal.component.model.ConfigurationMetadata;
 import org.gradle.internal.component.model.DefaultIvyArtifactName;
 import org.gradle.internal.component.model.DependencyMetadataRules;
 import org.gradle.internal.component.model.IvyArtifactName;
 import org.gradle.internal.component.model.ModuleSource;
+import org.gradle.internal.component.model.VariantMetadata;
 import org.gradle.internal.hash.HashUtil;
 import org.gradle.internal.hash.HashValue;
 import org.gradle.internal.reflect.Instantiator;
@@ -43,7 +50,7 @@ import java.util.Set;
 
 import static org.gradle.internal.component.model.ComponentResolveMetadata.DEFAULT_STATUS_SCHEME;
 
-abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultConfigurationMetadata> implements MutableModuleComponentResolveMetadata {
+abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultConfigurationMetadata> implements MutableModuleComponentResolveMetadata, ComponentVariantResolveMetadata {
     public static final HashValue EMPTY_CONTENT = HashUtil.createHash("", "MD5");
     private ModuleComponentIdentifier componentId;
     private ModuleVersionIdentifier id;
@@ -59,6 +66,11 @@ abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultCo
     private ImmutableMap<String, T> configurations;
 
     protected final Map<String, DependencyMetadataRules> dependencyMetadataRules = Maps.newHashMap();
+
+    private List<MutableVariantImpl> newVariants;
+    private ImmutableList<? extends ComponentVariant> variants;
+    private ImmutableList<? extends ConfigurationMetadata> graphVariants;
+
 
     protected AbstractMutableModuleComponentResolveMetadata(ModuleVersionIdentifier id, ModuleComponentIdentifier componentIdentifier, List<? extends ModuleDependencyMetadata> dependencies) {
         this.componentId = componentIdentifier;
@@ -77,6 +89,8 @@ abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultCo
         this.artifactOverrides = metadata.getArtifactOverrides();
         this.dependencies = metadata.getDependencies();
         this.contentHash = metadata.getContentHash();
+        this.variants = metadata.getVariants();
+        this.graphVariants = metadata.getVariantsForGraphTraversal();
     }
 
     @Override
@@ -232,6 +246,7 @@ abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultCo
         }
         dependencyMetadataRules.get(variantName).addAction(action);
         resetConfigurations();
+        graphVariants = null;
     }
 
     @Nullable
@@ -255,5 +270,204 @@ abstract class AbstractMutableModuleComponentResolveMetadata<T extends DefaultCo
     public void setDependencies(Iterable<? extends ModuleDependencyMetadata> dependencies) {
         this.dependencies = ImmutableList.copyOf(dependencies);
         resetConfigurations();
+    }
+
+    public MutableComponentVariant addVariant(String variantName, ImmutableAttributes attributes) {
+        MutableVariantImpl variant = new MutableVariantImpl(variantName, attributes);
+        if (newVariants == null) {
+            newVariants = new ArrayList<MutableVariantImpl>();
+        }
+        newVariants.add(variant);
+        graphVariants = null;
+        return variant;
+    }
+
+    public ImmutableList<? extends ConfigurationMetadata> getVariantsForGraphTraversal() {
+        if (graphVariants == null) {
+            ImmutableList<? extends ComponentVariant> variants = getVariants();
+            if (variants.isEmpty()) {
+                graphVariants = ImmutableList.of();
+            } else {
+                List<VariantBackedConfigurationMetadata> configurations = new ArrayList<VariantBackedConfigurationMetadata>(variants.size());
+                for (ComponentVariant variant : variants) {
+                    configurations.add(new VariantBackedConfigurationMetadata(getComponentId(), variant, dependencyMetadataRules.get(variant.getName())));
+                }
+                graphVariants = ImmutableList.copyOf(configurations);
+            }
+        }
+        return graphVariants;
+    }
+
+    public ImmutableList<? extends ComponentVariant> getVariants() {
+        if (variants == null && newVariants == null) {
+            return ImmutableList.of();
+        }
+        if (variants != null && newVariants == null) {
+            return variants;
+        }
+        ImmutableList.Builder<ComponentVariant> builder = new ImmutableList.Builder<ComponentVariant>();
+        if (variants != null) {
+            builder.addAll(variants);
+        }
+        for (MutableVariantImpl variant : newVariants) {
+            builder.add(new ImmutableVariantImpl(getComponentId(), variant.name, variant.attributes, ImmutableList.copyOf(variant.dependencies), ImmutableList.copyOf(variant.files)));
+        }
+        return builder.build();
+    }
+
+    @Override
+    public boolean definesVariant(String name) {
+        if (explicitlyDefinesVariants()) {
+            return containsNamedVariant(name);
+        } else {
+            return getConfigurationDefinitions().containsKey(name);
+        }
+    }
+
+    private boolean explicitlyDefinesVariants() {
+        return (variants != null && !variants.isEmpty()) || (newVariants != null && !newVariants.isEmpty());
+    }
+
+    private boolean containsNamedVariant(String name) {
+        if (variants != null) {
+            for (ComponentVariant variant : variants) {
+                if (variant.getName().equals(name)) {
+                    return true;
+                }
+            }
+        }
+        if (newVariants != null) {
+            for (MutableVariantImpl variant : newVariants) {
+                if (variant.getName().equals(name)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+
+    protected static class MutableVariantImpl implements MutableComponentVariant {
+        private final String name;
+        private final ImmutableAttributes attributes;
+        private final List<DependencyImpl> dependencies = new ArrayList<DependencyImpl>();
+        private final List<FileImpl> files = new ArrayList<FileImpl>();
+
+        MutableVariantImpl(String name, ImmutableAttributes attributes) {
+            this.name = name;
+            this.attributes = attributes;
+        }
+
+        @Override
+        public void addDependency(String group, String module, VersionConstraint versionConstraint) {
+            dependencies.add(new DependencyImpl(group, module, versionConstraint));
+        }
+
+        @Override
+        public void addFile(String name, String uri) {
+            files.add(new FileImpl(name, uri));
+        }
+
+        public String getName() {
+            return name;
+        }
+    }
+
+    protected static class FileImpl implements ComponentVariant.File {
+        private final String name;
+        private final String uri;
+
+        FileImpl(String name, String uri) {
+            this.name = name;
+            this.uri = uri;
+        }
+
+        @Override
+        public String getName() {
+            return name;
+        }
+
+        @Override
+        public String getUri() {
+            return uri;
+        }
+    }
+
+    protected static class DependencyImpl implements ComponentVariant.Dependency {
+        private final String group;
+        private final String module;
+        private final VersionConstraint versionConstraint;
+
+        DependencyImpl(String group, String module, VersionConstraint versionConstraint) {
+            this.group = group;
+            this.module = module;
+            this.versionConstraint = versionConstraint;
+        }
+
+        @Override
+        public String getGroup() {
+            return group;
+        }
+
+        @Override
+        public String getModule() {
+            return module;
+        }
+
+        @Override
+        public VersionConstraint getVersionConstraint() {
+            return versionConstraint;
+        }
+    }
+
+    protected static class ImmutableVariantImpl implements ComponentVariant, VariantMetadata {
+        private final ModuleComponentIdentifier componentId;
+        private final String name;
+        private final ImmutableAttributes attributes;
+        private final ImmutableList<DependencyImpl> dependencies;
+        private final ImmutableList<FileImpl> files;
+
+        ImmutableVariantImpl(ModuleComponentIdentifier componentId, String name, ImmutableAttributes attributes, ImmutableList<DependencyImpl> dependencies, ImmutableList<FileImpl> files) {
+            this.componentId = componentId;
+            this.name = name;
+            this.attributes = attributes;
+            this.dependencies = dependencies;
+            this.files = files;
+        }
+
+        @Override
+        public String getName() {
+            return name;
+        }
+
+        @Override
+        public DisplayName asDescribable() {
+            return Describables.of(componentId, "variant", name);
+        }
+
+        @Override
+        public ImmutableAttributes getAttributes() {
+            return attributes;
+        }
+
+        @Override
+        public ImmutableList<? extends Dependency> getDependencies() {
+            return dependencies;
+        }
+
+        @Override
+        public ImmutableList<? extends File> getFiles() {
+            return files;
+        }
+
+        @Override
+        public List<? extends ComponentArtifactMetadata> getArtifacts() {
+            List<ComponentArtifactMetadata> artifacts = new ArrayList<ComponentArtifactMetadata>(files.size());
+            for (ComponentVariant.File file : files) {
+                artifacts.add(new UrlBackedArtifactMetadata(componentId, file.getName(), file.getUri()));
+            }
+            return artifacts;
+        }
+
     }
 }
