+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/DefaultMutableMavenModuleResolveMetadata.java
@@ -17,27 +17,14 @@
 package org.gradle.internal.component.external.model;
 
 import com.google.common.collect.ImmutableList;
-import org.gradle.api.Action;
-import org.gradle.api.artifacts.DependenciesMetadata;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
-import org.gradle.api.artifacts.VersionConstraint;
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
 import org.gradle.api.internal.artifacts.ivyservice.ivyresolve.parser.GradlePomModuleDescriptorBuilder;
-import org.gradle.api.internal.attributes.ImmutableAttributes;
-import org.gradle.internal.Describables;
-import org.gradle.internal.DisplayName;
 import org.gradle.internal.component.external.descriptor.Configuration;
-import org.gradle.internal.component.model.ComponentArtifactMetadata;
-import org.gradle.internal.component.model.ConfigurationMetadata;
 import org.gradle.internal.component.model.DefaultIvyArtifactName;
-import org.gradle.internal.component.model.VariantMetadata;
-import org.gradle.internal.reflect.Instantiator;
-import org.gradle.internal.typeconversion.NotationParser;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 
 import static org.gradle.internal.component.external.model.DefaultMavenModuleResolveMetadata.JAR_PACKAGINGS;
@@ -47,9 +34,6 @@ public class DefaultMutableMavenModuleResolveMetadata extends AbstractMutableMod
     private String packaging = "jar";
     private boolean relocated;
     private String snapshotTimestamp;
-    private List<MutableVariantImpl> newVariants;
-    private ImmutableList<? extends ComponentVariant> variants;
-    private ImmutableList<? extends ConfigurationMetadata> graphVariants;
 
     /**
      * Creates default metadata for a Maven module with no POM.
@@ -73,8 +57,6 @@ public class DefaultMutableMavenModuleResolveMetadata extends AbstractMutableMod
         this.packaging = metadata.getPackaging();
         this.relocated = metadata.isRelocated();
         this.snapshotTimestamp = metadata.getSnapshotTimestamp();
-        variants = metadata.getVariants();
-        graphVariants = metadata.getVariantsForGraphTraversal();
     }
 
     @Override
@@ -100,37 +82,6 @@ public class DefaultMutableMavenModuleResolveMetadata extends AbstractMutableMod
     @Override
     protected Map<String, Configuration> getConfigurationDefinitions() {
         return GradlePomModuleDescriptorBuilder.MAVEN2_CONFIGURATIONS;
-    }
-
-    @Override
-    public boolean definesVariant(String name) {
-        if (explicitlyDefinesVariants()) {
-            return containsNamedVariant(name);
-        } else {
-            return getConfigurationDefinitions().containsKey(name);
-        }
-    }
-
-    private boolean explicitlyDefinesVariants() {
-        return (variants != null && !variants.isEmpty()) || (newVariants != null && !newVariants.isEmpty());
-    }
-
-    private boolean containsNamedVariant(String name) {
-        if (variants != null) {
-            for (ComponentVariant variant : variants) {
-                if (variant.getName().equals(name)) {
-                    return true;
-                }
-            }
-        }
-        if (newVariants != null) {
-            for (MutableVariantImpl variant : newVariants) {
-                if (variant.name.equals(name)) {
-                    return true;
-                }
-            }
-        }
-        return false;
     }
 
     @Nullable
@@ -174,175 +125,4 @@ public class DefaultMutableMavenModuleResolveMetadata extends AbstractMutableMod
         return JAR_PACKAGINGS.contains(packaging);
     }
 
-    @Override
-    public void addDependencyMetadataRule(String variantName, Action<DependenciesMetadata> action, Instantiator instantiator, NotationParser<Object, org.gradle.api.artifacts.DependencyMetadata> dependencyNotationParser) {
-        super.addDependencyMetadataRule(variantName, action, instantiator, dependencyNotationParser);
-        graphVariants = null;
-    }
-
-    @Override
-    public MutableComponentVariant addVariant(String variantName, ImmutableAttributes attributes) {
-        MutableVariantImpl variant = new MutableVariantImpl(variantName, attributes);
-        if (newVariants == null) {
-            newVariants = new ArrayList<MutableVariantImpl>();
-        }
-        newVariants.add(variant);
-        graphVariants = null;
-        return variant;
-    }
-
-    @Override
-    public ImmutableList<? extends ConfigurationMetadata> getVariantsForGraphTraversal() {
-        if (graphVariants == null) {
-            ImmutableList<? extends ComponentVariant> variants = getVariants();
-            if (variants.isEmpty()) {
-                graphVariants = ImmutableList.of();
-            } else {
-                List<VariantBackedConfigurationMetadata> configurations = new ArrayList<VariantBackedConfigurationMetadata>(variants.size());
-                for (ComponentVariant variant : variants) {
-                    configurations.add(new VariantBackedConfigurationMetadata(getComponentId(), variant, dependencyMetadataRules.get(variant.getName())));
-                }
-                graphVariants = ImmutableList.copyOf(configurations);
-            }
-        }
-        return graphVariants;
-    }
-
-    @Override
-    public ImmutableList<? extends ComponentVariant> getVariants() {
-        if (variants == null && newVariants == null) {
-            return ImmutableList.of();
-        }
-        if (variants != null && newVariants == null) {
-            return variants;
-        }
-        ImmutableList.Builder<ComponentVariant> builder = new ImmutableList.Builder<ComponentVariant>();
-        if (variants != null) {
-            builder.addAll(variants);
-        }
-        for (MutableVariantImpl variant : newVariants) {
-            builder.add(new ImmutableVariantImpl(getComponentId(), variant.name, variant.attributes, ImmutableList.copyOf(variant.dependencies), ImmutableList.copyOf(variant.files)));
-        }
-        return builder.build();
-    }
-
-    private static class MutableVariantImpl implements MutableComponentVariant {
-        private final String name;
-        private final ImmutableAttributes attributes;
-        private final List<DependencyImpl> dependencies = new ArrayList<DependencyImpl>();
-        private final List<FileImpl> files = new ArrayList<FileImpl>();
-
-        MutableVariantImpl(String name, ImmutableAttributes attributes) {
-            this.name = name;
-            this.attributes = attributes;
-        }
-
-        @Override
-        public void addDependency(String group, String module, VersionConstraint versionConstraint) {
-            dependencies.add(new DependencyImpl(group, module, versionConstraint));
-        }
-
-        @Override
-        public void addFile(String name, String uri) {
-            files.add(new FileImpl(name, uri));
-        }
-    }
-
-    private static class FileImpl implements ComponentVariant.File {
-        private final String name;
-        private final String uri;
-
-        FileImpl(String name, String uri) {
-            this.name = name;
-            this.uri = uri;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public String getUri() {
-            return uri;
-        }
-    }
-
-    private static class DependencyImpl implements ComponentVariant.Dependency {
-        private final String group;
-        private final String module;
-        private final VersionConstraint versionConstraint;
-
-        DependencyImpl(String group, String module, VersionConstraint versionConstraint) {
-            this.group = group;
-            this.module = module;
-            this.versionConstraint = versionConstraint;
-        }
-
-        @Override
-        public String getGroup() {
-            return group;
-        }
-
-        @Override
-        public String getModule() {
-            return module;
-        }
-
-        @Override
-        public VersionConstraint getVersionConstraint() {
-            return versionConstraint;
-        }
-    }
-
-    private static class ImmutableVariantImpl implements ComponentVariant, VariantMetadata {
-        private final ModuleComponentIdentifier componentId;
-        private final String name;
-        private final ImmutableAttributes attributes;
-        private final ImmutableList<DependencyImpl> dependencies;
-        private final ImmutableList<FileImpl> files;
-
-        ImmutableVariantImpl(ModuleComponentIdentifier componentId, String name, ImmutableAttributes attributes, ImmutableList<DependencyImpl> dependencies, ImmutableList<FileImpl> files) {
-            this.componentId = componentId;
-            this.name = name;
-            this.attributes = attributes;
-            this.dependencies = dependencies;
-            this.files = files;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public DisplayName asDescribable() {
-            return Describables.of(componentId, "variant", name);
-        }
-
-        @Override
-        public ImmutableAttributes getAttributes() {
-            return attributes;
-        }
-
-        @Override
-        public ImmutableList<? extends Dependency> getDependencies() {
-            return dependencies;
-        }
-
-        @Override
-        public ImmutableList<? extends File> getFiles() {
-            return files;
-        }
-
-        @Override
-        public List<? extends ComponentArtifactMetadata> getArtifacts() {
-            List<ComponentArtifactMetadata> artifacts = new ArrayList<ComponentArtifactMetadata>(files.size());
-            for (ComponentVariant.File file : files) {
-                artifacts.add(new UrlBackedArtifactMetadata(componentId, file.getName(), file.getUri()));
-            }
-            return artifacts;
-        }
-
-    }
 }
