+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/external/model/MavenModuleResolveMetadata.java
@@ -16,8 +16,6 @@
 
 package org.gradle.internal.component.external.model;
 
-import com.google.common.collect.ImmutableList;
-import org.gradle.internal.component.model.ConfigurationMetadata;
 import org.gradle.internal.component.model.ModuleSource;
 
 /**
@@ -42,6 +40,4 @@ public interface MavenModuleResolveMetadata extends ModuleComponentResolveMetada
     boolean isKnownJarPackaging();
     String getSnapshotTimestamp();
 
-    @Override
-    ImmutableList<? extends ConfigurationMetadata> getVariantsForGraphTraversal();
 }
